+++ Assets/Scripts/Companions/CompanionAbilityInstance.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4987cfad2002c46eb888e1852d8f5795
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
